[FIX-2] Refactor code, remove redundant logic
